Expand mercenary profile ID system from 255 to 2048 (Phases 1-7)
Widens the profile ID system throughout the engine while preserving full
backward compatibility with existing saves, .map files, and vanilla-range
(0-255) content.

- Phase 1-2: strongly-typed ProfileID (Soldier Control.h), call sites migrated
- Phase 3: NO_PROFILE sentinel fixed (0xFFFF), Profile Constants.h introduced
  to break a circular include
- Phase 4: NUM_PROFILES raised 255 -> 2048, gated by PROFILE_2048_EXPANSION
- Phase 5: map-file and editor compatibility
- Phase 6: remaining narrow profile-adjacent fields widened (VEHICLETYPE,
  PLAYERGROUP, MERCPROFILESTRUCT::bBuddy/bHated/bLearnToLike/bLearnToHate,
  bMercOpinion grown 255 -> 2048), each behind its own save-version gate
  with frozen legacy structs for old-format reads where the on-disk layout
  changed
- Phase 7: regression pass against a real pre-Phase-6 v1.13 save and a fresh
  new game, both through a local test environment (Data/ junctioned to a
  real v1.13 install, never modified). Found and fixed four genuine bugs in
  MERCPROFILESTRUCT::Load() surfaced by real-file testing: a buffer/alignment
  accounting gap in the old-format field reads, a missing version gate that
  caused a raw over-read of usDynamicOpinionFlagmask, two out-of-bounds reads
  in older conversion tiers, and a checksum hard-fail that aborted the whole
  load over one empty profile slot instead of using the caller's existing
  fallback. Confirmed working: 2048-profile save/load round-trip, sector
  entry / .map loading, Laptop UI, normal new-game flow.

See Tactical/Profile Constants.h for the full per-item changelog and known
remaining items (Rebel Command's bit-packed mercProfileId, Laptop/mercs.cpp's
hiring-availability subsystem, and the dynamic-opinion table's own resize
to 2048, queued as a follow-up).

+++ Ja2/Init.cpp
@@ -1320,7 +1320,7 @@ if( g_lang != i18n::Lang::en ) {
 }
 
 	// Old AIM Archive
-		UINT8 p;
+		UINT16 p;
 		for(p=0; p<NUM_PROFILES; p++)
 		{
 			gAimOldArchives[p].FaceID = -1;
